Corrige bug de deteccao de tasks existentes e adiciona quadrantes, horas editaveis e dark mode
- Corrige bug real: a chamada workitemsbatch usava "expand" em vez de
  "$expand" no corpo da requisicao, entao a Azure DevOps API ignorava
  o campo e nunca retornava relations - AlreadyHasTasks nunca era
  detectado corretamente com dados reais.
- Troca horas de int pra double em toda a cadeia de geracao de tasks
  (calculo, overrides, criacao) - o valor padrao de 4/8/16/24/40h
  sempre foi inteiro, mas a task real pode ser bem menor (ex: 0.1h).
- Work items da sprint: separa a listagem em dois quadrantes
  ("Demandas sem task" / "Demandas com task"), e adiciona campo de
  horas editavel na previa de cada task antes de gerar.
- Tasks por item: remove o toggle de mostrar/esconder tasks (fica
  sempre visivel), mesmas colunas da tela de tasks por parent, horas
  editaveis na previa, e botao de fechar cada task individualmente.
- Adiciona dark mode: toggle Claro/Escuro na tela de Configuracao,
  persistido em localStorage.

+++ backend/GestorDeBacklogs.Api/Services/IAzureDevOpsClient.cs
@@ -17,7 +17,7 @@ public interface IAzureDevOpsClient
 
     Task<Dictionary<string, JsonElement>> GetWorkItemRawFieldsAsync(int id, CancellationToken ct = default);
 
-    Task<int> CreateTaskAsync(WorkItemDto parent, string title, int hours, string iterationPath, CancellationToken ct = default);
+    Task<int> CreateTaskAsync(WorkItemDto parent, string title, double hours, string iterationPath, CancellationToken ct = default);
 
     Task CloseWorkItemAsync(int id, CancellationToken ct = default);
 }
